Revert "Add apex_defaults for consumerir HAL APEX"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6b147428-901b-431f-9883-f2b5aeeb2ea9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifb834a303911025ab33f267afd595911b7f21782