Somtetimes usb camera preview black
Sometimes v4l2 mjpg stream can't be decoded to yuv frame.
so the output thread call processCaptureRequestError() and go to
next loop.

But it forget to call waitForBufferRequestDone(). It will lead
issue of "BufferRequestThread does not support more than one concurrent request!"
in next loop. Also the requested buffers in current loop are not return
back to framework.

Change-Id: I25d437deda749c905d7f6c881214731b0cc10853
Signed-off-by: Fang Hui <hui.fang@nxp.com>