Sometimes usb camera preview black
If v4l2 mjpg stream failed to decode to yuv frame,
the output thread call processCaptureRequestError() and go to
next loop.

But it forgets to call waitForBufferRequestDone(). It will lead
issue of "BufferRequestThread does not support more than one concurrent request!"
in next loop. Also the requested buffers in current loop are not return
back to framework.

Change-Id: I25d437deda749c905d7f6c881214731b0cc10853
Signed-off-by: Fang Hui <hui.fang@nxp.com>