Revert "Ensure AIMapper and underlying IMPL outlive IMapperProvider"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/92878f09-9f3b-4a5f-89db-9f8ce64b730b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie58dfe744a8c4571ada052a30b8bac08092dc402