Revert "Cleanup: use shared apex_key (wifi HAL APEX)"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e0e9e80b-a0c1-4ee8-97af-dd9d4cb65cbb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I88b0be26fea01fe7559afbca3c11bce39423c65a