Revert "Update VtsHalEvsV1_*TargetTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/385347c8-71cf-4761-a033-76616d8367f5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0266c5886b15d2141f171a86b67778054825ad44