Revert "Cleanup: use shared apex_key (fingerprint APEX)"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/893aebdc-09ff-4247-ba83-5828a0d46370). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3bee6a2cc6b419c2836b0e4d64822ed837e0f4b6