Add remote submix direct paths for iec61937 wrapped eac3-joc
A new e-ac3-joc mix port pair is added for remote submix input/output.
This is to resolve DeviceHalAidl::findOrCreatePortConfig errors as part
of createAudioPatch where the patch format received (originating from
AudioPolicyManager) is the raw audio format, not the audio format opened
by AudioFlinger after SPDIF adaptation.

Skip endpoint match compatibility checking for cases where the source or
sink audio format is not PCM; this is to allow for the setAudioPatch
requests from AudioPolicyManager (from startInput -> setInputDevice)
with e-ac3-joc audio formats (not aware of the Spdif adapter in
AudioFlinger) causing patch failures; a better solution is for APM to
be made aware of the Spdif AF format adaptation, but this solution is
deferred.

Explicitly reject remote submix non PCM input or output streams. The
e-ac3-joc ports are added for compatibility with the current APM
createAudioPatch incorrect audio format due to AudioFlinger Spdif
adapters, but this should not lead to successfully created streams,
because it is required for the non-PCM streams to fail so that
AudioFlinger retries / reopens with PCM (wrapped IEC61937).

Change-Id: I3da9282d0f4ecb91dea65f784294e94436765538