Revert "Edit tests to adjust with changes in libxml2 upgrade"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3daeb529-47f6-41ee-bbad-8bf6c5fcd9da). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I897378d93eb055934e7f58e2284b2c6ca1f4f8e0