Revert "Remove Q matrices"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/77403502-9e75-48d2-926f-9348bf2d9453). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I48f61b24d793489c29493cc0d6fbe6c51ee8abda