Secretkeeper: add AuthGraph key exchange
Add `ISecretkeeper::getAuthGraphKE()` method to the Secretkeeper HAL.

Align the AIDL targets between AuthGraph and Secretkeeper, and add
some defaults that automatically link to the current version of the
Secretkeeper AIDL targets.

Move the non-secure implementation of AuthGraph to run the TA in a
separate thread.

Alter the nonsecure implementation of Secretkeeper so that TA-specific
code uses the `authgraph_comm::ta` code, running in a separate thread.
Also plug in an implementation of the AuthGraph TA code in this thread.
(However, the negotiated session keys are not yet used.)<F2>

Extend the Secretkeeper VTS tests to invoke the AuthGraph VTS inner
tests on the returned IAuthGraphKeyExchange instance, exercising the
instance as an AuthGraph sink.

Bug: 291228560
Test: VtsSecretkeeperTargetTest
Change-Id: Ia2c97976edc4530b2c902d95a74f3c340d342174