[WIP] Secretkeeper: add AuthGraph key exchange
Add `ISecretkeeper::getAuthGraphKE()` method to Secretkeeper.

Align the AIDL targets between AuthGraph and Secretkeeper.

Extend the nonsecure implementation of Secretkeeper to plug in a
nonsecure implementation of AuthGraph.  However, the negotiated
session keys are not yet used.

Extend the Secretkeeper VTS tests to invoke the AuthGraph VTS inner
tests on the returned IAuthGraphKeyExchange instance, exercising the
sink role.

Bug: 291228560
Test: VtsSecretkeeperTargetTest
Change-Id: Ia2c97976edc4530b2c902d95a74f3c340d342174