[Proposed change] Panic if the nonsecure instance is not found.
Since this test is specifically targeted for the nonsecure instance,
fail the test if the instance is not found.

Additionally, update the expected error code in corrupt_key test for
finish. In finish, we first check whether the own key exchange key is in
the opened key exchange sessions and fail if it is no with
INVALID_KE_KEY error.

Test: atest VtsAidlAuthGraphRoleTest
Change-Id: I4cfb408a2dde3f6ca28e89ef1346054ead7fcb83