Revert "Default implementation of getLeAudioBroadcastConfigurati..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6d69e450-5d08-4b67-8efa-da22e5a4acc2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9d04f0a0ee0a094d9da722e208e75abd52454d92