Clarify expected audio device description for callback
Device type BUS are used heavily in automotive and considered
as permanently attached. In audio framework, this means that
the connection type must be empty for such device types.

Updat the comment section to clarify that the expected
connection type when using IN/OUT_DEVICE is empty.

To resolve any ambiguity, starting from Android 15, there shall
be a new device type exclusively for bus: IN/OUT_BUS

Bug: 314832503
Test: Successful build

Change-Id: I1f3dac6d7276ad3a8b3ff6ce81e9dfbe8c210171