Revert "audio: Fix remote submix behavior with multiple inputs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6b2a2b8d-d50e-492a-a77a-7f1235777342). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I09f769eb81dd110e252cfec2e3d45073d65f912c