Add IStreamInEventCallback
External volume commands may be received from an audio input stream
source, such as a hardware volume control request to be applied in
software (in cases where no hardware volume is supported), or volume
requests from audio output streams which are looped back to the audio
input stream in the remote submix case. Add a new IStreamInEventCallback
to include volume changed notifications which are distinct from current
HW gain: the former is a request for volume to be applied by the capture
client, whereas HW gain is the gain applied in HW already.

Bug: 322260933
Change-Id: I0cb9a0d12c3b720093ec8dfae4d50e0bd381c387