Revert "audio: Add retries for BT proxy port registration"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c1240589-e2a0-4611-b878-649e8f0b8f66). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If134041e840d7c1b7d62c2e6b7e8ae7604b0aba5