Revert "audio: Move stable libaudioclient tests to presubmit for..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3633be69-6a6a-41d3-aea6-12ae7ef71389). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iba4077f28d8a56cdb9bf4c533d25e099ed2d1b44