Revert "Install bluetooth audio VINTF fragment in apex"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3b92dc3d-2a8e-4531-ab52-3c9f4990b875). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I85dbd7ddb71671258f9639a91d8a5a3aac81c12d