Revert "[DO NOT MERGE] vts: Fix for setProp in VtsHalAutomotiveV..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0838c315-b5ab-4de5-97ee-58809c9d7dfb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I22f125cf54d0270773e219dd79809c981c703009