Revert "Install bluetooth audio VINTF fragment in apex"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/632fa3c3-1563-4373-8668-89e16b376262). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If2946aed1fa7d5accaf255912add8d773c74d76c