Revert "Install bluetooth audio VINTF fragment in apex"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5eb1e009-365a-46f9-9aca-734bffadd67d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6cc47ae408d661105d058444c87a73e69fc2f89b