Revert "Install bluetooth audio VINTF fragment in apex"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/78ac15dc-6a3d-4905-94bc-084b9f7e8e30). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8eec5278fbe0d9e1597d0f8a240a123082cb7ec2