automotive aidl: Add required @RustDerive annotations
Derive vehicle hal parcelables from the Clone trait where applicable. This is is
supportive for implementing a vehicle hal service in Rust because caching e.g
VehiclePropValue or VehiclePropConfig instances allow less verbose code and
increase performance. Cloning a VehiclePropConfig instead of constructing each
time needed is less verbose. Property configs are normally not updated at
runtime.

This is not an API change, so in order to avoid having to bump the AIDL
API version for this HAL we used the hash_gen.sh script as below:

$ m android.hardware.automotive.vehicle-update-api
$ m android.hardware.automotive.vehicle-freeze-api

$ vim vehicle/aidl/Android.bp # removed frozen_api argument and the newly
created version 3 block

$ cp -r aidl_api/android.hardware.automotive.vehicle/3/* aidl_api/android.hardware.automotive.vehicle/2/
$ rm -rf aidl_api/android.hardware.automotive.vehicle/3/

./system/tools/aidl/build/hash_gen.sh \
    hardware/interfaces/automotive/vehicle/aidl/aidl_api/android.hardware.automotive.vehicle/2 \
    1 \
    hardware/interfaces/automotive/vehicle/aidl/aidl_api/android.hardware.automotive.vehicle/2/.hash

Change-Id: I586b89b160eb7a01ce24fad211b9e909175cdd3c