audio: Add a version of audio HAL VAPEX without preprocessing
The test `AudioEffectTest#ManageSourceDefaultEffects` fails on
CF automotive because software hotword detector occupies
the primary input. This test used to be skipped on HIDL HAL
because it did not have preprocessing effects. This CL adds
an audio HAL VAPEX configuration which also does not have
preprocessing effects, for use on automotive CF.

Bug: 318423731
Test: run `atest audioeffect_tests` on cf_x86_64_auto-trunk_staging-userdebug
Change-Id: I0f4ee9a44a3426934f6a055fc8c9ce74a8db78fc