Revert "Volume Control: Add tests to validate Volume Control Effect"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b851ab90-edfc-461e-a13f-cd86d3e9ea41). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ice75b140a204f27bcabff44c160987d7ca457909