Revert "uwb/aidl/default: Log incoming and outgoing packet bytes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1e97019c-719d-4017-b148-53293a83375a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I63b366f2a7d111199ec2f08fe967e3be03f3d97d