Revert "uwb/aidl/default: Use write_all instead of write in send..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1e97019c-719d-4017-b148-53293a83375a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7a7da843f6716906d6f6b1ef89ff2575b99a27e5