ExternalCameraHAL: Skip importing buffer from capture request
ExternalCameraHAL supports HAL buffer management which means
cameraservice will not send it an output buffer along with the
capture request, and the HAL has the freedom to request an
output buffer when an output buffer is needed.

As a remnant of migration from HIDL to AIDL, the
ExternalCameraHAL still attempted to import buffers that
the cameraservice has sent with the CaptureRequest. However,
with HAL buffer manager enabled, this buffer is always null and
results in the HAL failing to process the capture request.

This CL removes the logic for importing output buffers when
processing capture requests from the cameraservice, and lets
the HAL call requestStreamBuffers when it needs an output buffer.

Bug: 299182874
Test: VTS Tests now pass
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:43549d934ea08427d21401dabc0f3067fe2bf5bb)
Merged-In: I00654836b7ae91a91a2afa4b149712977e07dcc5
Change-Id: I00654836b7ae91a91a2afa4b149712977e07dcc5

NOTE FOR REVIEWERS - errors occurred while applying the patch.
PLEASE REVIEW CAREFULLY.
Errors:
Error applying patch in camera/device/default/ExternalCameraDeviceSession.cpp, hunk HunkHeader[1351,58->1357,6]: Hunk cannot be applied

Original patch:
 From 43549d934ea08427d21401dabc0f3067fe2bf5bb Mon Sep 17 00:00:00 2001
From: Avichal Rakesh <arakesh@google.com>
Date: Thu, 25 Jan 2024 16:08:52 -0800
Subject: [PATCH] ExternalCameraHAL: Skip importing buffer from capture request

ExternalCameraHAL supports HAL buffer management which means
cameraservice will not send it an output buffer along with the
capture request, and the HAL has the freedom to request an
output buffer when an output buffer is needed.

As a remnant of migration from HIDL to AIDL, the
ExternalCameraHAL still attempted to import buffers that
the cameraservice has sent with the CaptureRequest. However,
with HAL buffer manager enabled, this buffer is always null and
results in the HAL failing to process the capture request.

This CL removes the logic for importing output buffers when
processing capture requests from the cameraservice, and lets
the HAL call requestStreamBuffers when it needs an output buffer.

Bug: 299182874
Test: VTS Tests now pass
Change-Id: I00654836b7ae91a91a2afa4b1497129
[[[Original patch trimmed due to size. Decoded string size: 1037. Decoded string SHA1: e099a9791d793311a940955631be1cf842a1eeee.]]]

Change-Id: I0f42282cfa33f2cfd82d328897380d4136a5bdc3