Handle EVS extended information properly
setExtendedInfo() and getExtendedInfo() are methods that allow our
partners to implement custom commands / metadata. Because id and value
of extended information are opaque to the EVS service, behaviors of EVS
API handling the extended information need to be verified conditionally;
for example, when a test succeeds to set a value, it must be able to
read it back with the same id.

Bug: 329373218
Test: atest VtsHalEvsTargetTest
Change-Id: I965447abca8d1986057dc4a850d8af3f901aaf29