Revert "Fix failures in VtsHalRadioV1_5TargetTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bd54f86a-a4d7-48e9-90ca-85f98efe532e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 340355797
Change-Id: I0c60c9c118f8eb462bfde371d1a0b59e76b22510
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com