Revert "Fix failures in VtsHalRadioV1_5TargetTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0f3bc9c1-339c-493d-ab01-54cb76accf83). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 340355451
Change-Id: If460ec2aafe42072cbcb923165d0a0ebec3d00a5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com