Revert "Fix failures in VtsHalRadioV1_5TargetTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/63360eb8-8511-42d1-9732-497debf7d2ab). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 340355172
Change-Id: I3216185baccfce59c9efdcc4cc4a19e22b461e19
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com