Revert "Fix failures in VtsHalRadioV1_5TargetTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e92a7d9c-82c9-4322-b6fa-a9a9b290c8f6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 340355676
Change-Id: I2439a9db2194c002803865e7c2202283f1094c53
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com