Revert "Fix failures in VtsHalRadioV1_5TargetTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/398ac06d-7a9d-44ca-9cf6-7e96248ad7d8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 340355636
Change-Id: I9c1134a7df91af40eada693e6f4f994c2eb49ca4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com