Revert "Thread: simulate radio over Ethernet on Cuttlefish"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f1d80bb9-8e0c-4e9b-bda0-131f6f1fbd58). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 341213561
Change-Id: Iaf3cb71340f4bc295c316b94a304ffe04d27f391
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com