Revert "Fix test of triggerEmergencyNetworkScan"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0269c391-c309-47fd-be4c-2e0ea6758180). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7df88f4e0ae3a481206168558a6e2730c81d3e3b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com