Fix VtsHalSensorsV2_0TargetTest
During VTS testing, when many sensors are quickly enabled, the load on the Sensorhub side will be too heavy, and then continue to send sensor enable instructions and Sensor Flush instructions. These instructions sent by the Kernel driver to the Sensorhub through inter-core communication cannot be responded to in a timely manner. After the Kernel side judges the timeout, it will determine that the enable fails; after the enable fails, it cannot continue to test the flush or return the expected flush data.

Enable the sensors one by one to test whether they can run normally.

Bug: 339763843
Change-Id: I962f1c7080bb7cf63edf22912ab24ba16158bdcc
Test: run vts -m VtsHalSensorsV2_0TargetTest