Revert "[uwb-hal] add antenna mode app config param"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/21eec634-16b1-4fc1-bbf9-ea84998b548b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idf9189c3b03f7b75bcc47e45ce524c6d71b533f1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com