Revert "Remove Vndk_use_version property from AIDL"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a73d90d0-fe20-4a52-b7db-b0a0c9b71e45). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 345063864
Change-Id: Ied03d387f5f68f8492e50b96a47ac4d7db407372
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com