KeyMint: test attestation of imported keys
The KeyMint `importKey` API supports the generation of an attestation
for an imported key (by specifying `ATTESTATION_CHALLENGE` and
`ATTESTATION_APPLICATION_ID`), so add a test for this.

However, the test needs to deal with devices that are RKP-only and which
can therefore only generate attestations when an `ATTEST_KEY` is
provided.

Convert the existing code that copes with this situation (for
`GenerateKey`) into a helper method and re-use it.

Test: VtsAidlKeyMintTargetTest
Bug: 345187430
Change-Id: I2f9cc96b6f515102c7e7465228e3a5bb953a3471