Fix VtsHalSensorsV2_0TargetTest
During VTS testing, when many sensors are quickly enabled, the load on the Sensorhub side will be too heavy, and then continue to send sensor enable instructions and Sensor Flush instructions. These instructions sent by the Kernel driver to the Sensorhub through inter-core communication cannot be responded to in a timely manner. After the Kernel side judges the timeout, it will determine that the enable fails; after the enable fails, it cannot continue to test the flush or return the expected flush data.

ten sensors is tested as a group

Bug: 339763843
Test: run vts -m VtsHalSensorsV2_0TargetTest
Change-Id: I6f6d36d1e3c98b85b412189c3f97163c0945a7ab