Revert "Fix VtsHalSensorsV2_0TargetTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/81eb8e01-ce04-4b6e-93cd-af2c8c3e6d1c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I02af4816c53f94b75e4a955ddf78415bad67567e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com