Amend tests on GSI that rely on RKP-only props
GSI replaces the values for remote_prov_prop properties (since they’re
system_internal_prop properties), so on GSI the properties are not
reliable indicators of whether StrongBox/TEE are RKP-only or not.

Also included is the removal of the helper skipAttestKeyTestIfNeeded()
so the skipping can happen in the tests directly.

Bug: 348159232
Test: atest VtsAidlKeyMintTargetTest
Change-Id: I2075e1f76ddd0f87620a212e1aa389803139a117