Revert "Updated fuzz_config in Android.bp file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0235424f-4577-402d-b17e-b1248c148673). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1699383722d4e6f3021451315f92a454793164a9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com