KeyMint: default impl respects rkp-only property
If the system running the default/nonsecure KeyMint implementation is
marked as RKP-only, configure the KeyMint TA to be RKP-only.

DO NOT SUBMIT: doesn't work unless the HAL service has access to the
relevant property, but that's a system_internal property.  Re-visit
when there is a vendor property describing whether the device is
RKP-only.

Test: VtsAidlKeyMintTargetTest (with rkp_only = n and y)
Change-Id: I8403c37b0ec12d4486660ad88ba99ebcef5e3d1b