Revert "Restore Default Device Effects support with AIDL AudioHAL"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/44ab9472-42a2-410b-b92b-a04b8b87dbe1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iba8425da0ec8dfe9d22f1aadf29a5665ebd509e0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com