Revert "Optimize matching logic for multiple requirements at once."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/495dd928-8c94-4c92-b9a5-f3de6d961b66). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibcbd823851ae5171956216684b7695ac840d4573
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com