Revert "Update getDiceChainKind for VSR >15"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bad25ea0-a216-456b-8c31-0ac15a783f31). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie8d66a99c7e85ef4741e16a8f6b92c10990d39a5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com