Revert "Re-add the wildcard instance name for ISharedSecret in F..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6e50e492-8969-43c2-9d86-d6878a002ebd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 363085010
Change-Id: I2312de11d48175677c1447985fe7516de0227d40
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com