Revert "Define vintf_fragments as modules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/43dc9c87-bdb2-44a1-a784-e21255896ae7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie540028fbbc237e845f9ee835c4bc95023cfc4c9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com