Revert "Define vintf_fragments as modules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a0cee09b-a5d7-4aef-a9ab-85cbc035f379). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 363194678, 363194200, 363194759
Change-Id: I6f35132a171c635c235f60f6f89df8e72cd14f79
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com