Revert "Define vintf_fragments as modules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a3f8a7d2-3262-42e6-be0c-28cfc418c593). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 363194200, 363194678, 363194759
Change-Id: I9f9dc50c9629c46ce834a275390fb572a40231a8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com