Revert "Migrate forced config and for force use to android.media..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/021d68a5-9dc8-479f-ae1f-3db3a025c717). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie004e9d80321b06e3321e0c2d75c9a5ba10245d4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com