Revert "Migrate forced config and for force use to android.media..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/885d9f09-9c3c-4f23-82e0-d6c6ab325c49). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I380786c36e3aad56923a0c372ad2f06fffedf7c4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com