Convert to android.hardware.configstore@1.1-service
Add surfaceflinger_cc_defaults type to reflect
$(LOCAL_PATH)/surfaceflinger.mk included by
android.hardware.configstore@1.1-service. Skip the original require
condition which will not built configstore@1.1.policy in a coverage
build but required it not currently not allowed to be added into the
nest structure for converage and the required target is actually just a
prebuit text file which should not be an effort built it in coverage
build.

Bug: 357015079
Test: m android.hardware.configstore@1.1-service
Change-Id: Ie49de449c81f8559e2579025c2a2cf851fd9c93c