Add new constructor for fakevhal
Add new construcotr for fake vhal in order to change host side vhal to
enable suspend to disk/suspend to ram.
Since init was called by the construcotr, and C++ constructors are
constructed with the super classes constructor first, member variables
and member functions are not available until after the constructor
finishes, overriding getS2rS2dConfig is not sufficient.

Bug: 365994148
Test: manual
Change-Id: If32f0f5d5e3ad7de534ae16a73fdafda6f4594ee