Revert "Migrate forced config and for force use to android.media..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f131d6db-58ad-4f6e-a6b9-27b2637d546b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0216eda4d0243452d98338005330c90fefa4b517
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com