Revert "Add missing Cap Settings aidl support and helper"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f131d6db-58ad-4f6e-a6b9-27b2637d546b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2b0723e5fa7d75687fb36110b370a90be1424a41
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com