Revert "Fixed vts tests cases which do not support factory attes..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/25537986-b2f7-4749-b2ba-254359df987c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2bbfc7fb939200d7b149762a10f123f9edf28ea1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com