Revert "Support for P256 curve in RKP for Strongbox"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/25537986-b2f7-4749-b2ba-254359df987c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibfb95df5c7e88d21587697237d28df57bef20c8e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com