Revert "Support 192kHz sample rate in AIDL remote submix"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a7519def-04b6-4cf6-9744-36f4ae9f1307). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 370533763
Change-Id: Ia0db6073b727a8bca744f826e9e7b336f3754b1b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com