Revert "Use read-write lock to guard property config."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/48d00745-759d-44fa-8baa-3ecbe2719f9a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia7ac603b7124dfb344433fdc2e09be27a2f05b2a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com