Revert "Separate hdcp levels into a base interface"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1b54f2f8-cc5a-4fd3-81c5-78c3b9292908). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 370758572
Change-Id: I94ef08329886f769a68244bfa4fdfdfa9a44ee44
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com