Revert "Separate hdcp levels into a base interface"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cdbb7a24-bbd7-4885-a1e2-ba38ce1166a1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 370758706
Change-Id: Ic628cb79fa91d7f3c4f4b828a51a6d5b84dfb40c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com