Revert "Add a new known hash to drm"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/323e9433-cdfa-4cd0-bf53-b6ad3b515236). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 370987372
Change-Id: I69e2c58cc4de388f6de81d0f872b26c3468695ba
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com