Revert "Add a new known hash to drm"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d74d95c8-10b3-472a-b480-ea37c729adf6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 370987475
Change-Id: I5262a1f6d6a0c216098898bd0d8306d2329e325f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com