Revert "Add a new known hash to drm"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ff62e2fc-09de-4102-b4dd-7f347945c702). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 370987269
Change-Id: Ida6df95682c802defe7e7e593135169137cc53c9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com