Revert "Add a new known hash to drm"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e5a92c68-93e9-4378-993e-6a2659c88e5a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic22a0f0ae13658f3b5be409fd1b46341057953b4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com