Revert "Add a new known hash to drm"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ed05b4e0-d198-4b17-b989-7e878fbcbbe7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 371093931
Change-Id: I8a1de1c4bcf69a37724cb7a908862e7cfd311b25
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com