Revert "Add a new known hash to drm"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b93743d0-c76e-4744-a50f-9f5817281d75). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 371239459
Change-Id: Ic698a58d312521ccc40b3b1c2d73ffb8e746ecfe
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com