Revert "Use read-write lock to guard property config."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3903e004-1571-41b8-95f6-579606ce5161). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6212e9cf09e6bbf98963699cfc4b9041ba56a2dd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com