Revert "audio: Clarify that MMap h/w position must be continuous"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7ef324dd-905e-425b-94d6-dfbc16bc7106). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1c57cee7325f74ba3f6ea8af404cf6449dccfd92
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com