Revert "audio: Do not use StreamSwitcher for StreamPrimary"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fee5ea69-71c0-4572-a706-271944e3f26f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5bad1e7d5cff7d6a051d71262b3ab6ec7609ded9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com