Revert "audio: Deprecate StreamSwitcher"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1366d32a-165c-4942-8467-f31a9a330b68). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 372733172
Change-Id: I75b5de1fad12bc8f666971c3c38021c8f0b54f02
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com