Revert "Migrate forced config and for force use to android.media..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/88c97947-442f-4b16-955f-3f84c21e01c4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7d3193191ed4d01c22ccca6e512f0dcf99f72952
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com