Revert "Add missing Cap Settings aidl support and helper"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/88c97947-442f-4b16-955f-3f84c21e01c4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0cf0ea8e906483288419426555b738535348aa93
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com