Revert "Migrate forced config and for force use to android.media..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2410ae45-ea15-47c2-9fa7-7668ff464364). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic9455d1da0b8e6f5afd6d80f76f28931384796d4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com