Revert "Add missing Cap Settings aidl support and helper"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2410ae45-ea15-47c2-9fa7-7668ff464364). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I641bf07b49a1c9fb8abcea3485f33a092f50e238
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com