Enforce equal UDS public key for RKPVM & KeyMint
Work In Progress...

Add a VTS test to enforce the requirement that if the remotely
provisioned component called the Android Virtualization Framework
("avf") is available, then the DICE certificate chains for the
Android Virtualization Framework (a.k.a. protected VM,
a.k.a. Remote Key Provisioning VM) and the primary KeyMint ("default")
instances have the same root (i.e. the Unique Device Secret public key)

Test: m VtsHalRemotelyProvisionedComponentTargetTest
Test: m libkeymint_remote_prov_support_test
Test: atest hardware/interfaces/security/keymint/support

Bug: 372843164

Change-Id: If44e7ba927c7903c4e4429707fb8cf5c034dedb3