bootctl: set on_unlinked
one call of setonunlinked is triggered for each call of linktodeath. To
suppress the warnings of not calling setonunlinked with a non-null
cookie, we should add this here. We don't actually want to do anything
with the cookie (the boot control client) that is passed to the death
recipient, since that is still alive.

test: th
Change-Id: I55d1df7e27bdb926923ef599e18fe79b5de49352