Revert "Use platform security domains in keymint/gatekeeper sepolicy"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3f1cf59f-4ec0-4243-a733-49ba4a89336c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6295abef5b4b8b6a1f6fc81b6100c336c5adc326
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com