Revert "Enable OEM_PAID and OEM_PRIVATE APN types"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/18c2f262-cc67-4b02-a91b-ab83fc0f49c4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 375241855
Change-Id: I1d2bf5ced86f79342682f99f14722b596bfe867e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com