Revert "Enable OEM_PAID and OEM_PRIVATE APN types"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9db55998-75b7-489c-b137-987ab6239969). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 375256533
Change-Id: Ia8d3ec03b2b5ed206b2e23ef201e78b60c8a2d67
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com