Revert "Stage OEM_PAID and OEM_PRIVATE ApnTypes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9db55998-75b7-489c-b137-987ab6239969). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 375256533
Change-Id: Ief59b03085a700a89bc2ad50b08f4e4511bcf9b9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com