Revert "Use platform security domains in keymint/gatekeeper sepolicy"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/424cbdb0-6f34-4c10-b400-bf491be918f4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 375300226
Change-Id: Iad0bd9c10a6a7f01217c47e49aa155b1f0aa0b56
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com