equal root keys (UDS pub) in RKP VM/KeyMint DICE
adds a test to VtsRemotelyProvisionedComponentTests
that enforces equality of the root key (which is
the UDS public key) in the DICE certificate chain
for the RKP VM ("avf") and the primary
Keymint ("default") remotely provisioned component
instances.

There was significant refactoring performed to avoid duplicated parsing
and validation that occurred when

1) verifying an AuthenticatedRequest and extracting the CsrPayload
2) determining that the DiceCertChain was not degenerate
3) extracting the root public key of a DiceCertChain in a Csr

Note that 3) was new functionality that was to be added as specified
by the corresponding Bug number (see below).

Csr, AuthenticatedRequest, DiceCertChain, and CsrPayload

hardware/interfaces/security/rkp/aidl/android/hardware/security/keymint/generateCertificateRequestV2.cddl

A Csr is a specific type of AuthenticatedRequest and a DiceCertChain
lives in an AuthenticatedRequest. While more refactoring could have been
done, making an AuthenticatedRequest class was a first step in making
the code more modular.

The functionality that takes advantage of the AuthenticatedRequest
class live in the methods:

1) AuthenticatedRequest::csrPayload // retrieves the csrPayload
2) AuthenticatedRequest::isProper // is the DiceCertChain not degenerate
3) AuthenticatedRequest::getUdsPubFromDiceCertChain

A unit test was added to test the new functionality provided by 3)

Bug: 372843164
Test: m libkeymint_remote_prov_support_test
Test: atest hardware/interfaces/security/keymint/support
Test: m VtsHalRemotelyProvisionedComponentTargetTest
Change-Id: I73c18b47b4bb1d4e839afa272b0512f97bca17a4