Revert "Next kernel config is 6.12"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cebbc52a-c085-45ed-bd52-43713362bf08). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iee658ae688008de36b357ede53900b35753d4477
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com