Revert "Revert "audio: Do not use StreamSwitcher for StreamRemot..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cebbc52a-c085-45ed-bd52-43713362bf08). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia59fab23d1d751082a897623cda8f1e2e45ada7e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com